+++ cli/cmd/neverlauncher/release_commands.go
@@ -597,6 +597,8 @@ func releaseArtifacts(ver string) []string {
 		"neverlauncher-purpur-bridge-" + ver + ".jar",
 		"neverlauncher-folia-bridge-" + ver + ".jar",
 		"neverlauncher-fabric-bridge-" + ver + ".jar",
+		"neverlauncher-forge-bridge-" + ver + ".jar",
+		"neverlauncher-neoforge-bridge-" + ver + ".jar",
 		"BRIDGE_RELEASE_ALLOWLIST.json",
 		"BRIDGE_PLUGIN_MANIFEST.json",
 		"SBOM.spdx.json",